+++ examples/data-sources/account_logs/example_1.tf
@@ -1,0 +1,4 @@
+data "ovh_account_logs" "audit_logs" {
+  log_type        = "audit"
+  subscription_id = "xxx"
+}