+++ examples/resources/account_logs/example_1.tf
@@ -1,0 +1,5 @@
+resource "ovh_account_logs" "audit_logs" {
+  log_type      = "audit"
+  stream_id     = "xxx"
+  kind          = "default"
+}