+++ ovh/provider.go
@@ -101,6 +101,7 @@ func Provider() *schema.Provider {
 		},
 
 		DataSourcesMap: map[string]*schema.Resource{
+			"ovh_account_logs": dataSourceAccountLogs(),
 			"ovh_cloud_project_capabilities_containerregistry":               dataSourceCloudProjectCapabilitiesContainerRegistry(),
 			"ovh_cloud_project_capabilities_containerregistry_filter":        dataSourceCloudProjectCapabilitiesContainerRegistryFilter(),
 			"ovh_cloud_project_containerregistry_ip_restrictions_management": dataSourceCloudProjectContainerRegistryIPRestrictionsManagement(),
@@ -215,6 +216,7 @@ func Provider() *schema.Provider {
 		},
 
 		ResourcesMap: map[string]*schema.Resource{
+			"ovh_account_logs":                                               resourceAccountLogs(),
 			"ovh_cloud_project":                                              resourceCloudProject(),
 			"ovh_cloud_project_containerregistry":                            resourceCloudProjectContainerRegistry(),
 			"ovh_cloud_project_containerregistry_oidc":                       resourceCloudProjectContainerRegistryOIDC(),